Move controller-code to probdiffeq.adaptive because they are always used together
