+++ docs/benchmarks/pleiades/run_pleiades.py
@@ -18,7 +18,7 @@ import numpy as np
 import scipy.integrate
 import tqdm
 
-from probdiffeq import adaptive, ivpsolve, ivpsolvers, timestep
+from probdiffeq import adaptive, ivpsolve, timestep
 from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated
 from probdiffeq.solvers.strategies import filters
@@ -110,7 +110,7 @@ def solver_probdiffeq(*, num_derivatives: int, correction_fun) -> Callable:
         ts0_or_ts1 = correction_fun(ode_order=2)
         strategy = filters.filter_adaptive(ibm, ts0_or_ts1)
         solver = calibrated.dynamic(strategy)
-        control = ivpsolvers.control_proportional_integral()
+        control = adaptive.control_proportional_integral()
         adaptive_solver = adaptive.adaptive(
             solver, atol=1e-3 * tol, rtol=tol, control=control
         )
